Remove the bdrck_config subcrate, moving its contents to the root.
This is the first step of merging all of the bdrck_* subcrates into a
single crate. This is a simpler way to manage this code, and we can
achieve the same sort of "minimal dependency" footprint the old subcrate
method was trying to achieve using features, which is a much more
natural way to do this.

The version of the new monolithic crate is max(old crate versions) + 1.