Totally refactor the logging module.
This removes the "CLI logger" module entirely. I've decided that the
idea of using logging for both both actual log messages as well as
printing e.g. command-line interface stuff to the terminal was
misguided, for several reasons.

While we're at it, restructure the "debug logger" a bit to be more
generic. The intent is to eventually use it for all logging, not just
"debug logging", as well as to have it support features analogous to
Rust's env_logger crate.