Remove C++ build infrastructure.
At this point, we no longer care about being able to build the C++
version of bdrck, the code is just being kept around in case the
old implementation is useful to inspect while building the new Rust
implementation. So, the build infrastructure is no longer useful.

Projects which still rely on the C++ implementation should pin this
repository to the "cpp-final" tag.