Modify AbstractKey to just use generic failure::Errors.
Using an impl-specified error type is okay, until we want to write a
function which returns one of several possibly kinds of keys. We can't
use "impl trait" because it is for a single concrete type, and
Box<AbstractKey> likewise doesn't work because we would need to
explicitly specify the error type (and different types of keys could use
different types).

This requires underlying key libraries to convert their errors to
failure::Errors, but this generally doesn't lose information, so is
fine.