Split some flag types out into a separate crate.
This is the first step to introducing codegen for command callbacks.
Rust won't let us have both proc-macro definitions and normal types /
functions in the same crate, apparently. So, we need a separate crate
for the codegen. So, the end goal is to have a dependency tree like
this:

bdrck::flags============>flags-codegen
            \                 ||
             \                \/
              \=========>flags-values