Refactor keystore.rs to separate "loading" and "opening".
The idea here is, we want to be able to examine the key store's contents
(for example, looking at wrapping key signatures) *as part of deciding
which key we use to open it*.

So, we want to do it in a several-phase process, instead of all-at-once.
Imagine something like this:

- Load the key store from disk.
- Notice our system has a Yubikey connected.
- Check the keystore's contents; was this Yubikey used for wrapping?
- If yes, unwrap using it. If no, prompt for a password instead.

This commit only gets us part way there. This is more to be done to
fully enable this sort of situation, and also to "clean up" the API.