Don't require a key to construct KeyStores.
This is useful, because we want to call the constructor when we're
opening an existing KeyStore. In that case, the process we want to
follow is more like:

1. Load the existing KeyStore.
2. Examine the wrapping keys it contains.
3. Pick the best key to try to unwrap it with.

If we have to pass in a key up front, we can't first load the KeyStore
and examine it. :)

A consequence of this is that a KeyStore instance may exist in an
"unpersistable" state (i.e., with no wrapping keys); so, modify the
implementation to deal with this case gracefully. Add some API surface
to query the state of a KeyStore instance.

Finally, simplify the constructors of DiskKeyStore. Now that the
signatures of `new`, `load`, and `load_or_new` are all the same (because
none of them include a key), we can collapse them all into a single,
simpler constructor.

Finally, update tests to match this new API.