Refactor and simplify the key wrapping API.
For simplification, one key thing we're doing away with is nested
wrapping. It isn't something we actually use, and it significantly
complicates the API. So, now, there's just a simple wrap and unwrap
operation.

Likewise, get rid of this Wrappable trait. It just makes things more
complicated, and if we don't support nested wrapping there isn't really
a reason for it to exist.

Importantly, this lets us untangle a significant dependency that was
preventing adopting Secret more widely:

Keys no longer need to be Clone, Serialize, or Deserialize. Cloning
was a bad idea anyway, and with the simplified wrapping API we don't
need to do it anywhere. We do still need to "serialize" and
"deserialize" keys, but instead of doing it with serde, we define our
own API based around Secret.

NOTE: This changes the on-disk format of KeyStore!