Fixup crate initialization.
So halite-sys, and therefore crypto, is one example of a case where a
global initializer has to be used. Generally, we define a crate::init()
function which does this (depending on what features are enabled).

When switching away from sodiumoxide, we broke it. It compiles depending
on feature attributes, so we just stopped building the real version at
that point. We had:

    #[cfg(feature = "sodiumoxide")]

Annoyingly, when we removed the dependency entirely, this was just
false. So it silently broke. I don't know of a way to assert: "This
feature flag can be true or false, but it should certainly be a thing
that exists / is defined in Cargo.toml!", so this could happen again in
the future, but whatever.

So we fixed that.

To check for callers failing to do this, I added some debug assertions
which check that the library has been initialized, before calling into
FFI functions. This prevents undefined behavior.

It turns out, these assertions found some - our own tests!

Annoyingly, Rust doesn't seem to provide any way to do global test
initializers, without adding a bunch of dependencies. So, since it's
super easy to forget to add initialization to tests, I did "the next
best thing": I added a test, which looks at our test source code, and
checks that all of the tests have an initializer line.

And then, I fixed up the tests which this discovered. :)