Add more code preserving serde compatibility with older bdrck.
Previously, when we serialized Key, we did so with MessagePack, and Key
was a more complex structure. We've simplified it, which is great, but
we want to be able to deserialize keys which were serialized with old
versions of the code. And likewise, we want older version to be able to
interpret our output.

So, add some compatibility code which preserves the old format by adding
and stripping a small set of prefix bytes.